requirements.txt
riotwatcher
python-dotenv
discord